+++ packages/calcite-components/src/components/combobox/combobox.e2e.ts
@@ -28,12 +28,15 @@ import {
 } from "../../tests/utils/puppeteer";
 import { DEBOUNCE } from "../../utils/resources";
 import { ComponentTestTokens, themed } from "../../tests/commonTests/themed";
+import { mockConsole } from "../../tests/utils/logging";
 import { CSS } from "./resources";
 import { Combobox } from "./combobox";
 
 const selectionModes = ["single", "single-persist", "ancestors", "multiple"];
 
 describe("calcite-combobox", () => {
+  mockConsole();
+
   describe("renders", () => {
     renders("calcite-combobox", { display: "block" });
   });
@@ -752,12 +755,6 @@ describe("calcite-combobox", () => {
       </calcite-combobox>`,
     );
 
-    await page.waitForChanges();
-
-    const combobox = await page.find("calcite-combobox");
-    await combobox.callMethod("componentOnReady");
-    expect(combobox).not.toBeNull();
-
     const item = await page.find("calcite-combobox-item#item-0");
     let a11yItem = await page.find(`calcite-combobox >>> ul.${CSS.screenReadersOnly} li`);
 
@@ -800,7 +797,7 @@ describe("calcite-combobox", () => {
     item.setProperty("disabled", true);
     await page.waitForChanges();
     await page.waitForTimeout(DEBOUNCE.nextTick);
-    a11yItem = await page.find(`calcite-combobox >>> ul.${CSS.screenReadersOnly} li:nth-of-type(2)`);
+    a11yItem = await page.find(`calcite-combobox >>> ul.${CSS.screenReadersOnly} li`);
 
     expect(a11yItem).toBeNull();
   });
@@ -3018,7 +3015,6 @@ describe("calcite-combobox", () => {
           </calcite-combobox-item>
         </calcite-combobox>`,
       );
-      await page.waitForChanges();
     });
 
     async function testToggleAllItems(
@@ -3050,15 +3046,13 @@ describe("calcite-combobox", () => {
     }
 
     it("should toggle all items on and off with a click", async () => {
-      // eslint-disable-next-line @typescript-eslint/no-unused-vars
-      await testToggleAllItems(page, async ([selectAll, _combobox]) => {
+      await testToggleAllItems(page, async ([selectAll]) => {
         await selectAll.click();
       });
     });
 
     it("should toggle all items on and off with KeyDown press `enter`", async () => {
-      // eslint-disable-next-line @typescript-eslint/no-unused-vars
-      await testToggleAllItems(page, async ([_selectAll, combobox]) => {
+      await testToggleAllItems(page, async ([, combobox]) => {
         await combobox.press("Enter");
       });
     });
@@ -3100,7 +3094,6 @@ describe("calcite-combobox", () => {
       page: E2EPage,
       toggleAction: ([listItem, combobox]: [E2EElement, E2EElement]) => Promise<void>,
     ): Promise<void> {
-      const messages = await import("./assets/t9n/messages.json");
       const combobox = await page.find("calcite-combobox");
       await combobox.click();
       expect(await combobox.getProperty("open")).toBe(true);
@@ -3110,7 +3103,9 @@ describe("calcite-combobox", () => {
         item.setProperty("selected", true);
       }
       await page.waitForChanges();
-      expect(await page.find(`calcite-combobox >>> calcite-chip[value="${messages.allSelected}"]`)).toBeDefined();
+      expect(
+        await page.find(`calcite-combobox >>> calcite-chip[data-test-id="all-selected-indicator-chip"]`),
+      ).toBeDefined();
 
       const listItem = await combobox.find("calcite-combobox-item[value=Sequoia]");
       await toggleAction([listItem, combobox]);
@@ -3132,15 +3127,13 @@ describe("calcite-combobox", () => {
     }
 
     it("should toggle indeterminate state to `All Selected` when list items are toggled with a click", async () => {
-      // eslint-disable-next-line @typescript-eslint/no-unused-vars
-      await testToggleListItems(page, async ([listItem, _combobox]) => {
+      await testToggleListItems(page, async ([listItem]) => {
         await listItem.click();
       });
     });
 
     it("should toggle indeterminate state to `All Selected` when list items are toggled with a keydown `Enter`", async () => {
-      // eslint-disable-next-line @typescript-eslint/no-unused-vars
-      await testToggleAllItems(page, async ([_listItem, combobox]) => {
+      await testToggleAllItems(page, async ([, combobox]) => {
         await combobox.press("Enter");
       });
     });
@@ -3154,7 +3147,6 @@ describe("calcite-combobox", () => {
           </calcite-combobox-item>
         </calcite-combobox>`,
       );
-      await page.waitForChanges();
       const selectAll = await page.find(`calcite-combobox >>> calcite-combobox-item.${CSS.selectAll}`);
       expect(await selectAll.getProperty("indeterminate")).toBe(true);
     });
@@ -3168,7 +3160,7 @@ describe("calcite-combobox", () => {
           </calcite-combobox-item>
         </calcite-combobox>`,
       );
-      await page.waitForChanges();
+
       const selectAll = await page.find(`calcite-combobox >>> calcite-combobox-item.${CSS.selectAll}`);
       expect(await selectAll.getProperty("selected")).toBe(true);
     });
@@ -3183,8 +3175,6 @@ describe("calcite-combobox", () => {
           </calcite-combobox-item>
         </calcite-combobox>`,
       );
-      await page.waitForChanges();
-      const messages = await import("./assets/t9n/messages.json");
 
       const combobox = await page.find("calcite-combobox");
       await combobox.click();
@@ -3195,7 +3185,9 @@ describe("calcite-combobox", () => {
 
       expect(await page.find(`calcite-combobox >>> calcite-chip[value="Trees"]`)).toBeDefined();
       expect(await page.find(`calcite-combobox >>> calcite-chip[value="Maple"]`)).toBeDefined();
-      expect(await page.find(`calcite-combobox >>> calcite-chip[value="${messages.allSelected}"]`)).toBeNull();
+      expect(
+        await page.find(`calcite-combobox >>> calcite-chip[data-test-id="all-selected-indicator-chip"]`),
+      ).toHaveClass(CSS.chipInvisible);
     });
 
     it("should update aria-selected on items when toggling 'Select All'", async () => {
@@ -3215,7 +3207,7 @@ describe("calcite-combobox", () => {
       await selectAll.click();
       await page.waitForChanges();
 
-      a11yItem = await page.find(`calcite-combobox >>> ul.${CSS.screenReadersOnly} li:nth-of-type(2)`);
+      a11yItem = await page.find(`calcite-combobox >>> ul.${CSS.screenReadersOnly} li`);
       expect(await a11yItem.getProperty("ariaSelected")).toBe("false");
     });
   });
@@ -3319,6 +3311,31 @@ describe("calcite-combobox", () => {
         },
       };
       themed(singleSelectComboboxHTML, comboboxTokens);
+    });
+
+    const comboboxSelectAllEnabledHTML = html`
+      <calcite-combobox select-all-enabled>
+          </calcite-combobox-item value="Pine" text-label="Pine">
+          </calcite-combobox-item value="Not Pine" text-label="Not Pine">
+        </calcite-combobox>
+    `;
+
+    describe("select-all-enabled", () => {
+      themed(comboboxSelectAllEnabledHTML, {
+        "--calcite-combobox-divider-color": {
+          shadowSelector: `.${CSS.selectAll}`,
+          targetProp: "borderBlockEndColor",
+        },
+      });
+    });
+
+    describe("deprecated", () => {
+      themed(comboboxSelectAllEnabledHTML, {
+        "--calcite-combobox-item-border-color": {
+          shadowSelector: `.${CSS.selectAll}`,
+          targetProp: "borderBlockEndColor",
+        },
+      });
     });
   });
 });