+++ packages/calcite-components/src/components/stepper/stepper.tsx
@@ -332,8 +332,7 @@ export class Stepper extends LitElement {
     const { items, currentActivePosition } = this;
 
     let newIndex = startIndex;
-
-    while (items[newIndex]?.disabled && this.layout !== "horizontal-single") {
+    while (newIndex >= 0 && newIndex < items.length && items[newIndex]?.disabled) {
       newIndex = newIndex + (direction === "previous" ? -1 : 1);
     }
 
@@ -443,8 +442,10 @@ export class Stepper extends LitElement {
     const isPositionStart = position === "start";
     const path = isPositionStart ? "chevron-left" : "chevron-right";
     const { currentActivePosition, multipleViewMode, layout } = this;
-    const totalItems = this.items.length;
     const id = `${this.guid}-${isPositionStart ? "start" : "end"}`;
+    const offset = isPositionStart ? -1 : 1;
+    const direction = isPositionStart ? "previous" : "next";
+    const disabled = this.getEnabledStepIndex(currentActivePosition + offset, direction) === null;
 
     return layout === "horizontal-single" && !multipleViewMode ? (
       <calcite-action
@@ -455,10 +456,7 @@ export class Stepper extends LitElement {
         }}
         compact={true}
         data-position={position}
-        disabled={
-          (currentActivePosition === 0 && isPositionStart) ||
-          (currentActivePosition === totalItems - 1 && !isPositionStart)
-        }
+        disabled={disabled}
         icon={path}
         iconFlipRtl={true}
         id={id}