+++ packages/calcite-components/vite.config.ts
@@ -105,7 +105,6 @@ export default defineConfig({
     include: runBrowserTests ? [browserTestMatch] : [allSpecAndE2ETestMatch],
     exclude: runBrowserTests ? undefined : [...defaultExclude, browserTestMatch],
     passWithNoTests: true,
-    setupFiles: ["src/tests/setupTests.ts"],
   },
   /*
    * While useLumina() pre-configures everything for you, you can still